backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.8
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
filelock==3.15.4
fsspec==2024.6.1
gunicorn==23.0.0
huggingface-hub==0.24.5
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
setuptools==72.1.0
sqlparse==0.5.1
sympy==1.13.1
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0